Temporarily disable minimum trade.
Right now if you set a min trade amount, it will still tell you to
make transactions, despite there not being enough funds.

I need to move the rebalancing actions into another data layer that
has a function like "computeTrades", and takes into account all
of the stocks.  This can then be displayed as a list of actions
to take.